Use regex to locate word being completed
Instead of using the code completion information returned from the
language service a regex is now used to find the word at the current
caret position. The regex used is taken from the JSON client for
VSCode and is what is passed into VSCode's setLanguageConfiguration
function as the wordPattern parameter.

Code completion inside double quotes is working better with this
change however when the completion is triggered the text after the
current caret position is not replaced as it is in VSCode, but
remains after the completion text has been inserted.